Rebase

+++ LuckParser/Models/ParseModels/BoonLog.cs
@@ -8,17 +8,17 @@ namespace LuckParser.Models.ParseModels
     public class BoonLog
     {
         // Fields
-        private int time = 0;
-        private int value = 0;
-        private int overstack = 0;
+        private long time = 0;
+        private long value = 0;
+        private long overstack = 0;
 
         // Constructor
-        public BoonLog(int time, int value)
+        public BoonLog(long time, long value)
         {
             this.time = time;
             this.value = value;
         }
-        public BoonLog(int time, int value,int overstack)
+        public BoonLog(long time, long value, long overstack)
         {
             this.time = time;
             this.value = value;
@@ -26,16 +26,16 @@ namespace LuckParser.Models.ParseModels
         }
 
         // Getters
-        public int getTime()
+        public long getTime()
         {
             return time;
         }
 
-        public int getValue()
+        public long getValue()
         {
             return value;
         }
-        public int getOverstack() {
+        public long getOverstack() {
             return overstack;
         }
     }
