+++ LuckParser/Models/ParseModels/Duration.cs
@@ -15,13 +15,13 @@ namespace LuckParser.Models.ParseModels
 
         // Public Methods
         
-    public override int getStackValue()
+    public override long getStackValue()
         {
             // return boon_stack.stream().mapToInt(Integer::intValue).sum();
             //check for overflow
-            int total = boon_stack[0];
+            long total = boon_stack[0];
             for (int i =0;i<boon_stack.Count-1;i++) {
-                if (total > 0 && boon_stack[i + 1] > int.MaxValue - total)
+                if (total > 0 && boon_stack[i + 1] > long.MaxValue - total)
                 {
                     //Overflow
                     return int.MaxValue;
@@ -35,13 +35,13 @@ namespace LuckParser.Models.ParseModels
         }
 
         
-    public override void update(int time_passed)
+    public override void update(long time_passed)
         {
 
             if (boon_stack.Count() > 0)
             {
                 // Clear stack
-                if (time_passed >= getStackValue())
+                if (time_passed >= (long)getStackValue())
                 {
                     boon_stack.Clear();
                     return;
@@ -49,11 +49,11 @@ namespace LuckParser.Models.ParseModels
                 // Remove from the longest duration
                 else
                 {
-                    boon_stack[0] = (boon_stack[0] - time_passed);
+                    boon_stack[0] = (int)((long)boon_stack[0] - time_passed);
                     if (boon_stack[0] <= 0)
                     {
                         // Spend leftover time
-                        time_passed = Math.Abs(boon_stack[0]);
+                        time_passed = (long)Math.Abs(boon_stack[0]);
                         boon_stack.RemoveAt(0);
                         update(time_passed);
                     }
@@ -62,7 +62,7 @@ namespace LuckParser.Models.ParseModels
         }
 
         
-    public override void addStacksBetween(List<int> boon_stacks, int time_between)
+    public override void addStacksBetween(List<long> boon_stacks, long time_between)
         {
         }
     }
