+++ LuckParser/Models/ParseModels/MechanicLog.cs
@@ -9,14 +9,14 @@ namespace LuckParser.Models.ParseModels
     public class MechanicLog
     {
         // Fields
-        private int time;
+        private long time;
         private int skill_id;
         private string name;
         private int damage;
         private Player player;
         private string plotlyShape;
 
-        public MechanicLog(int time, int skill_id,string name, int damage, Player p,string plotlyShape) {
+        public MechanicLog(long time, int skill_id,string name, int damage, Player p,string plotlyShape) {
             this.time = time;
             this.skill_id = skill_id;
             this.name = name;
@@ -25,7 +25,7 @@ namespace LuckParser.Models.ParseModels
             this.plotlyShape = plotlyShape;
         }
         //getters
-        public int GetTime() {
+        public long GetTime() {
             return time;
         }
         public int GetSkill()
